refactor: move non-components to 'lib' dir
