+++ iconoir.com/components/Typography.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
-import { showNotification } from '../helpers/showNotification';
+import { showNotification } from '../lib/showNotification';
 import { Button } from './Button';
-import { media } from './responsive';
+import { media } from '../lib/responsive';
 
 export const Text15 = styled.div`
   font-size: 15px;
