+++ iconoir.com/pages/index.tsx
@@ -4,7 +4,7 @@ import type { NextPage } from 'next';
 import styled from 'styled-components';
 import { AvailableFor } from '../components/AvailableFor';
 import { LargeButton } from '../components/Button';
-import { REPO, SUPPORT_LINK } from '../components/constants';
+import { REPO, SUPPORT_LINK } from '../lib/constants';
 import { Explore } from '../components/Explore';
 import { Footer } from '../components/Footer';
 import { Header } from '../components/Header';
@@ -12,7 +12,7 @@ import { HeaderBackground } from '../components/HeaderBackground';
 import { Icon } from '../components/IconList';
 import { Layout } from '../components/Layout';
 import { Praise } from '../components/Praise';
-import { media } from '../components/responsive';
+import { media } from '../lib/responsive';
 import { SEO } from '../components/SEO';
 import { Stat, StatsContainer } from '../components/Stats';
 import { Text18 } from '../components/Typography';
