+++ iconoir.com/pages/support.tsx
@@ -4,12 +4,12 @@ import Image from 'next/image';
 import styled from 'styled-components';
 import { HeroText } from '.';
 import { LargeButton } from '../components/Button';
-import { DISCORD_LINK, SUPPORT_LINK } from '../components/constants';
+import { DISCORD_LINK, SUPPORT_LINK } from '../lib/constants';
 import { Footer } from '../components/Footer';
 import { Header, HeaderProps } from '../components/Header';
 import { HeaderBackground } from '../components/HeaderBackground';
 import { Layout } from '../components/Layout';
-import { media } from '../components/responsive';
+import { media } from '../lib/responsive';
 import { SEO } from '../components/SEO';
 import { Code, Heading2, Text18 } from '../components/Typography';
 import { getHeaderProps } from '../lib/getHeaderProps';
